prepare for navigation
